Function scale_decode::decode

pub fn decode<V: Visitor>(
    data: &mut &[u8], 
    ty_id: u32, 
    types: &PortableRegistry, 
    visitor: V
) -> Result<V::Value, V::Error>

Decode data according to the type ID and PortableRegistry provided. The provided pointer to the data slice will be moved forwards as needed depending on what was decoded, and a method on the provided Visitor will be called depending on the type that needs to be decoded.