scale_decode::visitor

Module types

Types used in the super::Visitor trait definition.

Structs

Array

This enables a visitor to decode items from an array type.

BitSequence

This represents a BitSequence, deferring decoding until the implementation wants to.

Composite

This represents a composite type.

Sequence

This enables a visitor to decode items from a sequence type.

Str

This represents a string, but defers proper decoding of it until it’s asked for, and avoids allocating.

Tuple

This represents a tuple of values.

Variant

A representation of the a variant type.