Module scale_decode::visitor

The Visitor trait and associated types.

Structs

Array

This represents an array type.

BitSequence

This represents a BitSequence, deferring decoding until the implementation wants to.

Compact

This represents a compact encoded type.

Composite

This represents a composite type.

IgnoreVisitor

A Visitor implementation that just ignores all of the bytes.

Sequence

This enables a visitor to decode items from a sequence type.

Str

This represents a string, but defers proper decoding of it until it’s asked for, and avoids allocating.

Tuple

This represents a tuple of values.

TypeId

The ID of the type being decoded.

Variant

A representation of the a variant type.

Enums

BitSequenceValue

A decoded BitVec.

CompactLocation

A pointer to what the compact value is contained within.

DecodeError

An error decoding SCALE bytes.

Traits

Visitor

An implementation of the Visitor trait can be passed to the crate::decode() function, and is handed back values as they are encountered. It’s up to the implementation to decide what to do with these values.